Add `database.CreateSchema()` and `DropSchema()`.
This change involves exposing the previously internal `database` package.

To make for a better public-facing API, the `New()` method has been changed to
return an error instead of panicking.

A `MustNew()` function has been added to retain the previous behavior of `New()`.
