+++ database/doc.go
@@ -1,0 +1,3 @@
+// Package database provides functions for initializating the SQLite database
+// used to store projection data.
+package database
