+++ projections/customer_test.go
@@ -3,7 +3,7 @@ package projections_test
 import (
 	"testing"
 
-	"github.com/dogmatiq/example/internal/database"
+	"github.com/dogmatiq/example/database"
 	"github.com/dogmatiq/example/internal/testrunner"
 	"github.com/dogmatiq/example/messages/commands"
 	"github.com/dogmatiq/testkit"
@@ -14,7 +14,7 @@ func Test_CustomerProjectionHandler(t *testing.T) {
 	t.Run(
 		"when an account is opened for a new customer",
 		func(t *testing.T) {
-			db := database.New()
+			db := database.MustNew()
 			defer db.Close()
 
 			testrunner.New(db).
